requirements.txt
streamlit
openai
python-docx
moviepy
